[Improve][PluginDiscovery] Move into seatunnel-core-starter module
