+++ .github/workflows/backend.yml
@@ -122,7 +122,6 @@ jobs:
               - "seatunnel-core/**"
               - "seatunnel-e2e/seatunnel-e2e-common/**"
               - "seatunnel-formats/**"
-              - "seatunnel-plugin-discovery/**"
               - "seatunnel-transforms-v2/**"
               - "seatunnel-translation/**"
               - "seatunnel-e2e/seatunnel-transforms-v2-e2e/**"
