+++ docs/en/contribution/coding-guide.md
@@ -21,7 +21,6 @@ This guide documents an overview of the current Apache SeaTunnel modules and bes
 | seatunnel-examples                     | SeaTunnel local examples module, developer can use it to do unit test and integration test                                                                         |
 | seatunnel-engine                       | SeaTunnel engine module, seatunnel-engine is a new computational engine developed by the SeaTunnel Community that focuses on data synchronization.                 |
 | seatunnel-formats                      | SeaTunnel formats module, used to offer the ability of formatting data                                                                                             |
-| seatunnel-plugin-discovery             | SeaTunnel plugin discovery module, used to offer the ability of loading SPI plugins from classpath                                                                 |
 | seatunnel-transforms-v2                | SeaTunnel transform V2 module, currently transform V2 is under development and the community will focus on it                                                      |
 | seatunnel-translation                  | SeaTunnel translation module, used to adapt Connector V2 and other computing engines such as Spark Flink etc...                                                    |
