+++ pom.xml
@@ -83,7 +83,6 @@
         <module>seatunnel-connectors-v2</module>
         <module>seatunnel-api</module>
         <module>seatunnel-translation</module>
-        <module>seatunnel-plugin-discovery</module>
         <module>seatunnel-formats</module>
         <module>seatunnel-engine</module>
         <module>seatunnel-examples</module>
