+++ seatunnel-core/seatunnel-core-starter/src/main/java/org/apache/seatunnel/core/starter/plugin/PluginIdentifier.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.seatunnel.plugin.discovery;
+package org.apache.seatunnel.core.starter.plugin;
 
 import org.apache.commons.lang3.StringUtils;
