+++ seatunnel-core/seatunnel-core-starter/src/test/resources/home/connectors/plugin-mapping.properties
