+++ seatunnel-core/seatunnel-spark-starter/src/main/java/org/apache/seatunnel/core/starter/spark/execution/TransformExecuteProcessor.java
@@ -22,8 +22,8 @@ import org.apache.seatunnel.api.table.type.SeaTunnelDataType;
 import org.apache.seatunnel.api.table.type.SeaTunnelRow;
 import org.apache.seatunnel.api.transform.SeaTunnelTransform;
 import org.apache.seatunnel.core.starter.exception.TaskExecuteException;
-import org.apache.seatunnel.plugin.discovery.PluginIdentifier;
-import org.apache.seatunnel.plugin.discovery.seatunnel.SeaTunnelTransformPluginDiscovery;
+import org.apache.seatunnel.core.starter.plugin.PluginIdentifier;
+import org.apache.seatunnel.core.starter.plugin.discovery.SeaTunnelTransformPluginDiscovery;
 import org.apache.seatunnel.translation.spark.common.serialization.InternalRowConverter;
 import org.apache.seatunnel.translation.spark.common.utils.TypeConverterUtils;
